Limit deploy to just this app
This pulls up the deploy job parameters and adds the ansible variable to
limit the deploy to just this app. This is useful because it stops the
issue where all sscs nodejs apps are always deployed when any one needs
to be deployed.

This change also pulls out the call to sscs-track-your-appeal-deploy
into a separate function so that it can be easily reused after the smoke
tests are run.

We want the dev environment to tend towards master on every app, so
after the smoke tests are run the master branch should be deployed.