Refactor notifucations to not use session
Refactor the notifications routes to pass the token around in the URL
rather than using a session. This makes testing easier and avoids the
complexities of using a session.

For calling the TokenService I've used the same pattern that's used for
calling the AppealService on the /progress routes. This adds a bit of
indirection to the routes but removes a lot of duplication.

I have rearranged the routes for stopping and changing notifications to
give the URLs a more obvious hierarchy /resource/identifier/action.