Refactor app middleware to route middleware
We use middleware to do actions common across a set of routes. This is
useful for removing duplication but using app middleware means we don't
have access to expressjs's parsing of the URLs. This change refactors
these into route level middleware so we can use `req.params`. As an
added benefit this means that we could now test these in isolation.