SSCS-708 Implement error pages
https://triadmoj.atlassian.net/browse/SSCS-708

Implement error pages for 404 and 500 errors.

Ideally 500 errors would be handled outside of the application but as we
do not currently have a way of doing that we handle them in the app and
accept that complete app failure won't be caught.

This change adds explicit routes for 404 and 500 pages so that the load
balancers have some way to retrieve the error pages.