Move isActive into nunjucks filter
Move the isActive funciton into a Nunjucks filter rather than the
context. This is a more obvious place to work with it as that is where a
lot of our other template functions live.

I've also introduced jsdom as a dependency so that we can more easily
test template output.