Move locale into res.locals
Moving the locale into the res.locals object makes it easier to factor
out parts of the application into libraries.

This change adds the locale to the res.locals object so that it can be
retrieved from the request context rather than globally. This change
does not select which locale to use it just makes a locale (at the
moment just english) available to the request.